cdk_examples/requirements.txt
aws-cdk-lib==2.135.0  
boto3==1.34.84
constructs>=10.3.0,<11.0.0